Refactor generator support into a dedicated module (for lowering)
